Rename point/sources.hpp to point/source.hpp
