+++ core/specfem/point.hpp
@@ -7,6 +7,6 @@
 #include "point/kernels.hpp"
 #include "point/partial_derivatives.hpp"
 #include "point/properties.hpp"
-#include "point/sources.hpp"
+#include "point/source.hpp"
 #include "point/stress.hpp"
 #include "point/stress_integrand.hpp"
